[tests] - move modeling rules to integration-tests
